Major Updates
User can pull to refresh
User can compose a new tweet by tapping on a compose button.
User can tap on a tweet to view it, with controls to retweet, favorite,
and reply.
User can retweet, favorite, and reply to the tweet directly from the
timeline feed.
Adds Counter for characters left when composing a Tweet (NavBar)
Retweeting and Favoriting Increment Counters and change icons
User can load more tweets once they reach the bottom of the feed using
infinite loading similar to the actual Twitter client.
After creating a new tweet, this is added to the timeline immediately
without refetching the timeline from the network.